# api/swagger.yaml
basePath: /api/v1
definitions:
  domain.OrgDTO:
    properties:
      name:
        type: string
    type: object
  respfmt.Fmt:
    properties:
      data:
        type: object
      developerMessage:
        type: string
      serviceName:
        type: string
      totalCount:
        type: integer
      userMessage:
        type: string
    type: object
info:
  contact:
    email: aleksandr.melentyev@nexign.com
    name: Aleksandr Melentyev
  description: API server for bizzen application
  title: bizzen application API
  version: "1.0"
paths:
  /org:
    get:
      consumes:
      - application/json
      operationId: read-all-org
      parameters:
      - description: limit
        in: query
        name: limit
        required: true
        type: integer
      - description: offset
        in: query
        name: offset
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/respfmt.Fmt'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/respfmt.Fmt'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/respfmt.Fmt'
      summary: Getting a organizations list
      tags:
      - org
    post:
      consumes:
      - application/json
      operationId: create-org
      parameters:
      - description: Organization data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.OrgDTO'
      produces:
      - application/json
      responses:
        "201":
          description: created
          schema:
            $ref: '#/definitions/respfmt.Fmt'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/respfmt.Fmt'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/respfmt.Fmt'
      summary: Creating a new organization
      tags:
      - org
  /org/{id}:
    delete:
      consumes:
      - application/json
      operationId: soft-delete-org-by-id
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/respfmt.Fmt'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/respfmt.Fmt'
        "404":
          description: not found
          schema:
            $ref: '#/definitions/respfmt.Fmt'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/respfmt.Fmt'
      summary: Filling deletion data
      tags:
      - org
    get:
      consumes:
      - application/json
      operationId: read-org-by-id
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/respfmt.Fmt'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/respfmt.Fmt'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/respfmt.Fmt'
      summary: Getting organization by ID
      tags:
      - org
    put:
      consumes:
      - application/json
      operationId: update-org-by-id
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: integer
      - description: organization data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.OrgDTO'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/respfmt.Fmt'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/respfmt.Fmt'
        "404":
          description: not found
          schema:
            $ref: '#/definitions/respfmt.Fmt'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/respfmt.Fmt'
      summary: Updating organization data
      tags:
      - org
  /org/{id}/history:
    get:
      consumes:
      - application/json
      operationId: read-org-history-by-id
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/respfmt.Fmt'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/respfmt.Fmt'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/respfmt.Fmt'
      summary: Getting organization row changes in table
      tags:
      - org
swagger: "2.0"